added shortcuts for orders (press Alt + A and the shortcut)
From the manual: To give an order with a shortcut, press Alt + A. A message will tell the list of available shortcuts for the currently controlled unit. Press the shortcut and the order will be executed immediately, unless a target needs to be selected.

From the modding manual: Simple orders, building orders, training orders, orders using an ability can be given with a shortcut, if a shortcut is defined. To define a shortcut, define a "shortcut" property followed by the corresponding letter. The letter must be in lowercase. If the order is a simple order, the shortcut must be defined by the order (ex: patrol). If the order is a complex order (train, build, use an ability), the shortcut must be defined by the second part of the order. For example, define an "m" shortcut for the meteor ability so the mage will have the "m" shortcut to cast meteors.