refactoring: translated cible to action_target; renamed the previous action_target to a private name: _action_target
"cible" could have been called "target", but the name is also used by Order, and more refactoring is planned, so it will be convenient to have different names for a while. Action and Unit.action might be added later, and Unit.target would be replaced with Unit.action.target. This makes sense because a unit have a list of orders to execute, and a current action. Although the current action might be linked to the current order, an action can also be the result of an initiative (a reaction to an event, for example the arrival of an enemy).