refactoring: renamed walk_to_xy() to go_to_xy(); now air units use go_to_xy() instead of action_fly_to_remote_target()
Air units still don't have a graph for the allowed exits of a square, so they move in a straight line from the start to the target. While this makes sense, it has the drawback to scatter units between several squares. If the units followed a list of squares, they would always be focused on one square at a time. So maybe later the air units will also have a graph of the possible paths (to model impassable mountains) and the air units will follow the shortest path.