first version of units blocking an exit
Now "block" is the default action when the target is an exit.
A unit won't move as long as it is blocking an exit.
A unit acts as a gate: it lets other units move if there is no enemy around.
At the moment one unit is enough to block any exit.