servers that support different mods (by Thiago Seus)
I've changed the server code so that it registers only the version in the
servers list, and the clients compare this version when connecting.
However, the clients still send a compatibility version string, and the
server isolates clients with different version strings.
This way, only players who use the same settings can invite each other.

Conflicts:
	soundrts/servermain.py
	soundrts/serverroom.py