restored the profiler
This is a first version of the return of the profiler. I initially thought that cProfile couldn't work with multithreads, but it is actually possible.

http://stackoverflow.com/questions/653419/how-can-i-profile-a-multithread-program-in-python
"Instead of running one cProfile, you could run separate cProfile instance in each thread, then combine the stats. Stats.add() does this automatically."

I was initially misguided by the question and its best reply. And the alternate tool (yappi) is difficult to install on Windows (no precompiled binaries).
Vartec's reply is much better, because I usually prefer to use the default tools if they are good enough. And now specific parts of the program can be profiled, so it's even better.