removed "Alt+Tab" bugfix
The bugfix is unnecessary since pygame 1.9.2.

The problem was with SDL (at least the one included in Pygame 1.9.1), at least on Windows 10: when focusing back with Alt+Tab, or when starting the game with a Control+Alt desktop shortcut, the modifier keys stick until they are pressed and released again.

The problem was mentioned there: https://github.com/pygame/pygame/issues/436