bug fixed: pressing Space will tell the exact orders even when some units have different orders
This is very useful to check how many workers are gathering gold, wood, etc (by pressing D).
This could be useful to know how many units in a group are moving and how many have arrived.
Pressing Control + Shift + S will give a complete summary of the orders of soldiers and workers.