pygame 2: restored the default behavior for non US-QWERTY keyboard layouts
With pygame 1, on Windows, using K_* constants and event.key, the keyboard layout was always US-QWERTY, allowing the easy definition of "physical" key bindings: only the position of the key mattered, not the final symbol of the key. On Linux, event.key used the keyboard layout of the local system, so the result was different.

With pygame 2, on Windows, event.key uses the keyboard layout of the local system, but it's apparently still possible to easily define "physical" keys by using undocumented KSCAN_* constants and event.scancode. If the behavior is the same on Linux and Mac, it will be an improvement on these platforms.

Note: from now on, pygame 2 is required.